+++ src/pages/workspaces/workspace/workflows/EntitySelectionType.js
@@ -2,5 +2,6 @@ export default {
   processAll: Symbol('process all'),
   processMergedSet: Symbol('process merged set'),
   chooseRows: Symbol('choose rows'),
-  chooseSets: Symbol('choose sets')
+  chooseSets: Symbol('choose sets'),
+  chooseSetComponents: Symbol('choose set components')
 }
